Pancake flips/ Pancake lands flat on Plate and sticks
Pancake flips via a rotate per minute deal activited by the slingshot code when the pancake is released to fire. Seems like their should be a better way to do this so that it stops rotation through physics when it hits anything.

As of now when the pancake hits the top of the plate it sets of a trigger set in the empty game object 'plate's' mesh collider.

The trigger stops the rotation of pancake, sets it's transfrom.rotation to 0,0,0, and makes the pancake a child of the plate. I still feel like the rotation and stopping the roation could be done better. But the child thing might help.

We might be able to create some sort of function in plate that makes the last landed (top) pancake take over the functionality of the trigger that makes the pancakes stay in place. That way we can stack them on top of eachother.

By the way you can't submit unity stuff on here without closing the unity window.